+++ src/Main/MainState.ts
@@ -1,3 +1,4 @@
 export interface MainState {
     porp: string;
+    logoutRequest: boolean;
 };